refactor(wds-mcp): 코드 리팩토링
