+++ packages/wds-mcp/src/constants/index.ts
@@ -1,0 +1,4 @@
+export const DOCS_BASE_URL = 'https://dev-montage.wanted.co.kr';
+
+export const DOCS_COLOR_USAGE_URL = `${DOCS_BASE_URL}/docs/foundations/base-material/colors/semantic`;
+export const DOCS_SELECTOR = '[data-algolia-page-scope="true"]';
