+++ packages/wds-mcp/src/helpers/index.ts
@@ -1,10 +1,34 @@
 import * as icons from '@wanteddev/wds-icon';
 import { theme } from '@wanteddev/wds-theme';
 import { load } from 'cheerio';
+import { camelCase, kebabCase } from 'change-case';
 
 import api from '../../../../docs/generated/api.json' assert { type: 'json' };
+import { DOCS_BASE_URL } from '../constants';
 
 const componentApi = api;
+
+export const getGuideUrls = async () => {
+  const sitemap = await fetch(`${DOCS_BASE_URL}/sitemap.xml`);
+  const sitemapXml = await sitemap.text();
+
+  const $sitemap = load(sitemapXml);
+  const urls = $sitemap('urlset url loc')
+    .map((i, el) => $sitemap(el).text())
+    .get();
+
+  return urls.filter((url) => {
+    if (url.includes('/docs/components') && url.endsWith('/web')) {
+      return true;
+    }
+
+    if (url.includes('/docs/utilities/web-')) {
+      return true;
+    }
+
+    return false;
+  });
+};
 
 export const listComponents = () => {
   const components = componentApi.map((component) => component.name).sort();
@@ -70,24 +94,65 @@ export const listTokens = () => {
   });
 };
 
-export const getGuideUrls = async () => {
-  const sitemap = await fetch('https://montage.wanted.co.kr/sitemap.xml');
-  const sitemapXml = await sitemap.text();
+export const listUtilityFunctions = async () => {
+  const guideUrls = await getGuideUrls();
 
-  const $sitemap = load(sitemapXml);
-  const urls = $sitemap('urlset url loc')
-    .map((i, el) => $sitemap(el).text())
-    .get();
+  const utilityUrls = guideUrls.filter(
+    (url) =>
+      url.includes('/web-utilities/') &&
+      !url.endsWith('/navigation') &&
+      !url.endsWith('/media'),
+  );
 
-  return urls.filter((url) => {
-    if (url.includes('/docs/components') && url.endsWith('/web')) {
-      return true;
-    }
+  return [
+    ...utilityUrls.map((value) => camelCase(value)),
+    'respondTo',
+    'respondDown',
+    'respondMore',
+    'respondUp',
+    'useMediaQuery',
+  ];
+};
 
-    if (url.includes('/docs/utilities/web-')) {
-      return true;
-    }
+export const getComponentUrl = async (componentName: string) => {
+  const componentSlug = kebabCase(componentName);
+  const componentPathMap: Record<string, string> = {
+    list: 'list-cell',
+    stepper: 'progress-tracker',
+    'card-list': 'card',
+    modal: 'popup',
+  };
+  const customComponentPath = componentPathMap[componentSlug];
 
-    return false;
-  });
+  const guideUrls = await getGuideUrls();
+
+  return (
+    guideUrls.find((url) =>
+      customComponentPath
+        ? url.endsWith(`${customComponentPath}/web`)
+        : url.endsWith(`${componentSlug}/web`),
+    ) ??
+    guideUrls.find((url) =>
+      customComponentPath
+        ? url.endsWith(`${customComponentPath}`)
+        : url.endsWith(`${componentSlug}`),
+    )
+  );
+};
+
+export const getUtilityFunctionUrl = async (functionName: string) => {
+  const utilityFunctionPathMap: Record<string, string> = {
+    respondTo: 'media',
+    respondDown: 'media',
+    respondMore: 'media',
+    respondUp: 'media',
+    useMediaQuery: 'media',
+  };
+  const customUtilityFunctionPath = utilityFunctionPathMap[functionName];
+
+  const guideUrls = await getGuideUrls();
+
+  return guideUrls.find((url) =>
+    url.endsWith(kebabCase(customUtilityFunctionPath ?? functionName)),
+  );
 };
